requirements.txt
streamlit
opencv-python
numpy
soundfile
librosa
pandas
pyttsx3
streamlit-webrtc
mediapipe
tensorflow
torch
fer
deepface
scikit-learn
pyaudio
SpeechRecognition
langchain
google-generativeai
python-dotenv
sentence-transformers
nltk
tqdm
scipy
matplotlib
protobuf==3.20.*